+++ src/components/ExpandingBlocks/index.scss
