+++ src/components/ExpandingBlocks/index.tsx
